Units and enemies parameters update
-Changes on units collider size
-Changes on enemies speed, health and damage
-Changes on projectiles speed and damage.

+++ Assets/Scripts/Objective.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Objective : MonoBehaviour
 {
-    public int health = 30;
+    public float health = 30;
 
     private void Awake()
     {
